requirement.txt
asgiref==3.7.2
Django==4.2.4
django-colorfield==0.9.0
django-tinymce==3.6.1
Pillow==10.0.0
python-slugify==8.0.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
